requirements.txt
torch>=2.1
transformers>=4.40
datasets>=2.18
sentence-transformers>=2.7
faiss-cpu>=1.8
numpy>=1.26
pyyaml>=6.0
